+++ Plugin.cs
@@ -22,7 +22,7 @@ namespace Tollbound
     {
         public const string PluginGuid = "com.ragemedia.tollbound";
         public const string PluginName = "Tollbound";
-        public const string PluginVersion = "0.2.1";
+        public const string PluginVersion = "0.2.2";
 
         internal static TollboundPlugin Instance;